Updated webpack config
* Fixed web-build delete
* Moved large functions out of config file
