+++ apps/native-component-list/package.json
@@ -8,7 +8,8 @@
   "scripts": {
     "postinstall": "expo-yarn-workspaces postinstall",
     "web:dev": "webpack-dev-server -d --open --config webpack/webpack.dev.js",
-    "web:build": "webpack -p --config webpack/webpack.prod.js",
+    "web:build": "webpack -p --optimize-minimize --config webpack/webpack.prod.js",
+    "web:build:stats": "webpack -p --optimize-minimize --config webpack/webpack.prod.js --json > web-stats.json",
     "web:publish": "netlify deploy --dir web-build",
     "web:serve": "serve ./web-build"
   },
@@ -31,6 +32,7 @@
     "processing-js": "^1.6.6",
     "prop-types": "^15.6.0",
     "react": "16.5.0",
+    "react-art": "^16.8.2",
     "react-dom": "16.7.0",
     "react-native": "0.57.1",
     "react-native-is-iphonex": "^1.0.1",
@@ -69,6 +71,7 @@
     "webpack": "4.24.0",
     "webpack-bundle-analyzer": "^3.0.4",
     "webpack-cli": "^3.2.3",
+    "webpack-deep-scope-plugin": "^1.6.0",
     "webpack-dev-server": "^3.2.0",
     "webpack-manifest-plugin": "^2.0.4",
     "webpack-merge": "^4.2.1",
