+++ apps/native-component-list/webpack/webpack.common.js
@@ -5,142 +5,10 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const ManifestPlugin = require('webpack-manifest-plugin');
 const WorkboxPlugin = require('workbox-webpack-plugin');
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
-
+const env = require('./getClientEnvironment');
 const locations = require('./webpackLocations');
 const nativeAppManifest = require(locations.appJson);
-
-function getAppManifest() {
-  if (nativeAppManifest && nativeAppManifest.expo) {
-    const { expo } = nativeAppManifest;
-    const PWAManifest = require(locations.template.manifest);
-    const web = PWAManifest || {};
-
-    return {
-      // facebookScheme
-      // facebookAppId
-      // facebookDisplayName
-      name: expo.name,
-      description: expo.description,
-      slug: expo.slug,
-      sdkVersion: expo.sdkVersion,
-      version: expo.version,
-      githubUrl: expo.githubUrl,
-      orientation: expo.orientation,
-      primaryColor: expo.primaryColor,
-      privacy: expo.privacy,
-      icon: expo.icon,
-      scheme: expo.scheme,
-      notification: expo.notification,
-      splash: expo.splash,
-      androidShowExponentNotificationInShellApp: expo.androidShowExponentNotificationInShellApp,
-      web,
-    };
-  }
-  return {};
-}
-const environment = process.env.NODE_ENV || 'development';
-const __DEV__ = environment !== 'production';
-
-const ENV_VAR_REGEX = /^(EXPO_|REACT_NATIVE_)/i;
-
-const publicUrl = '';
-
-function getClientEnvironment() {
-  let processEnv = Object.keys(process.env)
-    .filter(key => ENV_VAR_REGEX.test(key))
-    .reduce(
-      (env, key) => {
-        env[key] = JSON.stringify(process.env[key]);
-        return env;
-      },
-      {
-        // Useful for determining whether we’re running in production mode.
-        // Most importantly, it switches React into the correct mode.
-        NODE_ENV: JSON.stringify(environment),
-
-        // Useful for resolving the correct path to static assets in `public`.
-        // For example, <img src={process.env.PUBLIC_URL + '/img/logo.png'} />.
-        // This should only be used as an escape hatch. Normally you would put
-        // images into the root folder and `import` them in code to get their paths.
-        PUBLIC_URL: JSON.stringify(publicUrl),
-
-        // Surface the manifest for use in expo-constants
-        APP_MANIFEST: JSON.stringify(getAppManifest()),
-      }
-    );
-  return {
-    'process.env': processEnv,
-    __DEV__,
-  };
-}
-
-function generateHTMLFromAppJSON() {
-  const { expo: expoManifest = {} } = nativeAppManifest;
-
-  const metaTags = {
-    viewport: 'width=device-width, initial-scale=1, shrink-to-fit=no',
-    description: expoManifest.description || 'A Neat Expo App',
-    'theme-color': expoManifest.primaryColor || '#000000',
-    'apple-mobile-web-app-capable': 'yes',
-    // default, black, black-translucent
-    'apple-mobile-web-app-status-bar-style': 'default',
-    'apple-mobile-web-app-title': expoManifest.name,
-    'application-name': expoManifest.name,
-    // Windows
-    'msapplication-navbutton-color': '',
-    'msapplication-TileColor': '',
-    'msapplication-TileImage': '',
-  };
-
-  // Generates an `index.html` file with the <script> injected.
-  return new HtmlWebpackPlugin({
-    /**
-     * The file to write the HTML to.
-     * Default: `'index.html'`.
-     */
-    filename: locations.production.indexHtml,
-    /**
-     * The title to use for the generated HTML document.
-     * Default: `'Webpack App'`.
-     */
-    title: expoManifest.name,
-    /**
-     * Pass a html-minifier options object to minify the output.
-     * https://github.com/kangax/html-minifier#options-quick-reference
-     * Default: `false`.
-     */
-    minify: {
-      removeComments: true,
-      /* Prod */
-      collapseWhitespace: true,
-      removeRedundantAttributes: true,
-      useShortDoctype: true,
-      removeEmptyAttributes: true,
-      removeStyleLinkTypeAttributes: true,
-      keepClosingSlash: true,
-      minifyJS: true,
-      minifyCSS: true,
-      minifyURLs: true,
-    },
-    /**
-     * Adds the given favicon path to the output html.
-     * Default: `false`.
-     */
-    favicon: locations.template.favicon,
-    /**
-     * Allows to inject meta-tags, e.g. meta: `{viewport: 'width=device-width, initial-scale=1, shrink-to-fit=no'}`.
-     * Default: `{}`.
-     */
-    meta: metaTags,
-    /**
-     * The `webpack` require path to the template.
-     * @see https://github.com/jantimon/html-webpack-plugin/blob/master/docs/template-option.md
-     */
-    template: locations.template.indexHtml,
-  });
-}
-
-const env = getClientEnvironment();
+const indexHTML = require('./getIndexHTMLFromAppJSON');
 
 const includeModule = module => {
   return path.resolve(locations.modules, module);
@@ -246,7 +114,6 @@ function useWebModule(modulePathToHiJack, redirectPath, initialRoot = 'expo/buil
 const publicPath = '/';
 
 module.exports = {
-  // mode: environment,
   context: __dirname,
   // configures where the build ends up
   output: {
@@ -265,25 +132,12 @@ module.exports = {
     },
     runtimeChunk: 'single',
   },
-  module: {
-    // strictExportPresence: true,
-
-    rules: [
-      { parser: { requireEnsure: false } },
-
-      htmlLoaderConfiguration,
-      babelLoaderConfiguration,
-      imageLoaderConfiguration,
-      ttfLoaderConfiguration,
-      mediaLoaderConfiguration,
-    ],
-  },
   plugins: [
     // Generates an `index.html` file with the <script> injected.
-    generateHTMLFromAppJSON(),
+    indexHTML,
 
     new InterpolateHtmlPlugin(HtmlWebpackPlugin, {
-      PUBLIC_URL: publicUrl,
+      PUBLIC_URL: publicPath,
       WEB_TITLE: nativeAppManifest.expo.name,
     }),
 
@@ -299,9 +153,7 @@ module.exports = {
 
     useWebModule('Performance/Systrace', 'Performance/Systrace'),
     useWebModule('RCTNetworking', 'Network/RCTNetworking'),
-    /** Upstreamed */
     useWebModule('Platform', 'Utilities/Platform'),
-    /** Upstreamed */
     useWebModule('HMRLoadingView', 'Utilities/HMRLoadingView'),
 
     new WorkboxPlugin.GenerateSW({
@@ -309,7 +161,7 @@ module.exports = {
       clientsClaim: true,
       exclude: [/\.LICENSE$/, /\.map$/, /asset-manifest\.json$/],
       importWorkboxFrom: 'cdn',
-      navigateFallback: `${publicUrl}/index.html`,
+      navigateFallback: `${publicPath}index.html`,
       navigateFallbackBlacklist: [new RegExp('^/_'), new RegExp('/[^/]+\\.[^/]+$')],
       runtimeCaching: [
         {
@@ -320,52 +172,30 @@ module.exports = {
     }),
     new BundleAnalyzerPlugin({
       analyzerMode: 'static',
-      openAnalyzer: false,
+      // openAnalyzer: false,
     }),
   ],
+
+  module: {
+    // strictExportPresence: true,
+
+    rules: [
+      { parser: { requireEnsure: false } },
+
+      htmlLoaderConfiguration,
+      babelLoaderConfiguration,
+      imageLoaderConfiguration,
+      ttfLoaderConfiguration,
+      mediaLoaderConfiguration,
+    ],
+  },
   resolve: {
     symlinks: false,
     extensions: ['.web.js', '.js', '.jsx', '.json'],
-    alias: Object.assign(
-      {
-        /* Alias direct react-native imports to react-native-web */
-        'react-native$': 'react-native-web',
-        'react-dom$': require.resolve('react-dom'),
-        'react-art$': require.resolve('react-art'),
-      },
-      [
-        'ActivityIndicator',
-        'Alert',
-        'AsyncStorage',
-        'Button',
-        'DeviceInfo',
-        'Modal',
-        'NativeModules',
-        'Network',
-        'Platform',
-        'SafeAreaView',
-        'SectionList',
-        'StyleSheet',
-        'Switch',
-        'Text',
-        'TextInput',
-        'TouchableHighlight',
-        'TouchableWithoutFeedback',
-        'View',
-        'ViewPropTypes',
-      ].reduce(
-        (acc, curr) => {
-          acc[curr] = `react-native-web/dist/cjs/exports/${curr}`;
-          return acc;
-        },
-        {
-          JSEventLoopWatchdog: 'react-native-web/dist/cjs/vendor/react-native/JSEventLoopWatchdog',
-          React$: 'react',
-          ReactNative$: 'react-native-web/dist/cjs',
-          infoLog$: 'react-native-web/dist/cjs/vendor/react-native/infoLog',
-        }
-      )
-    ),
+    alias: {
+      /* Alias direct react-native imports to react-native-web */
+      'react-native$': 'react-native-web',
+    },
   },
   // Some libraries import Node modules but don't use them in the browser.
   // Tell Webpack to provide empty mocks for them so importing them works.
